added error handling and fixed prototype issues

+++ libft/ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: knieve-l <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/19 17:59:54 by knieve-l          #+#    #+#             */
-/*   Updated: 2024/09/19 19:29:03 by knieve-l         ###   ########.fr       */
+/*   Updated: 2024/09/20 11:25:48 by knieve-l         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,9 +20,9 @@ static int	ft_isspace(int c)
 
 int	ft_atoi(const char *str)
 {
-	int	i;
-	int	sign;
-	int	result;
+	int		i;
+	int		sign;
+	long	result;
 
 	i = 0;
 	sign = 1;
